+++ src/beta/easypost.js
@@ -1,3 +1,6 @@
+import BetaPaymentRefund, {
+  propTypes as betaPaymentRefundTypes,
+} from './resources/beta_payment_refund';
 import Referral, { propTypes as referralPropTypes } from './resources/referral';
 
 import API from '../easypost';
@@ -5,10 +8,12 @@ import API from '../easypost';
 export const DEFAULT_BASE_URL = 'https://api.easypost.com/beta/';
 
 export const RESOURCES = {
+  BetaPaymentRefund,
   Referral,
 };
 
 export const PROP_TYPES = {
+  betaPaymentRefundTypes,
   referralPropTypes,
 };